Upload Solidity artifacts and rust test-runner
During the release process we should update all the artifacts required
for downstream testing without focing them to re-build everything. In
this case the test-runner is a single binary that runs all the tests and
contains all the other functions inside. We also include the Peggy.json
solidity contract artifact so that users can always deploy Gravity
bridge as it was at a given release rather than having to deploy from
master.