Fix sequential event nonce check
This sneaky inline error caused the attestation to be processed based on
the correct validator address but the event nonce lookup was processed
on the delegate address instead of the validator address.

This didn't matter in the tests becuase events are submitted in batches
and we simply never ended up in a desync situation.

We should carefully inspect the codebase for more casts like this.