Add test for London hardfork
This patch adds a new test, specifically to test out thew new London
hardfork and how Gravity behaves with the new fee market.

This test is independent from the others rather than updating all tests
to use the London fork, becuase we need to maintain compatibility with
existing Eth semantics for the time being.

This test updates our test network Geth to a version that is compatible
with the London hardfork.