Direct testnet logs to disk
This change directs all logs for the Cosmos chain to the disk and sets
the log level from error to info, this should allow for improved
debugging with the tradeoff of additional diskspace being used. Also
errors may not be obvious by looking at the logs in CI since they are no
longer printed by default.

For now this is acceptible for ease of local debugging. We'll see if
this tradeoff requires revisiting.