Add "HARDHAT" option for all integration tests
This patch adds a HARDHAT option to allow all integration tests to
be easily run on OSX targets that behave poorly with Geth's mining
feature.

This also moves the eth chain start logic into a single file and adds
some comments.

We can't use Hardhat by default because it's not correct for several of
our tests.