Add more complex CancelSendToEth tests
This commit adds two new test conditions to transaction_stress_test one
is where a random user cancels all three of thier sends and we make sure
they do not execute.

The second is where this same user attempts to rudely cancel the
transaction of another user, which we ensure does *not* happen.

Is also patches a reliability issue with the transaction stress tests
centered around how gas prices change during the tests.