Hardcode power in Gravity.sol
The Gravity.sol contract has since it's inception taken the the number
of votes required to pass a measure as a constructor argument.

This hasn't been effective for quite some time, since that number become
effectively hardcoded in a few other places.

This patch takes that de-facto standard and makes it cannonical, the
power in the contract is now hardcoded and a new documentation section
has been added about how and where to change values to modify the voting
resolution or passing percentage.