Upgrade Solidity 0.6.6 -> 0.8.7
This patch updates to the latest solidity version, reducing security
vulnerabilities and taking advantage of new gas optimizations.

The most specific vulnerability resolved by this change has been that
previously not all operations used safemath.

We did not use safemath for nonce operations and this created the threat
of overflowing nonces being exploited.

This patch removes some of the arbitrary logic tests, as they depended
on UniswapV2 tooling which is no longer possible to import.