Use custom error messages in Gravity.sol
This patch uses custom Solidity error messages to reduce the total size
of the contract and assist in future debugging.

This patch has a pretty poor risk/reward ratio when it comes to having
ot invert all of our logic from fail on false to fail on true. The
benefits are not huge, but the risks are if DeMorgans law isn't properly
applied.

This patch does logically change the behavior of the one million nonce
increment lockout, to be one million and one blocks before it becomes
actually invalid. This should be the only logical change in this patch.