String deposits for Gravity.sol
This patch changes to sendToCosmos() method to use strings, rather than
a bytes32 array, as the encoding for a deposit.

Note that this breaks the existing API and all existing frontends.

The reason for doing this is to make it possible to interpret different
bech32 prefixes and forward them as required to different chains. It
could also be used in the future to indicate almost anything else we may
want.

The cost for this improvement is an enormous new amount of error
handling which must be tested to extreme detail. We must deal with every
possibility of an invalid deposit.

To that end we send deposits that are invalid for any reason into the
community pool on the Cosmos chain. The reasoning here is that it's
actually pretty hard to create an invalid deposit unintentionally,
espeically as a client using tested frontends.