Do not exit the relayer on GravityID failure
Becuase we used 'return' instead of 'continue' here the orchestrator would unexpectedly stop if at any time during it's loops it encountered a bad response form the Ethereum node.

Since the relayer is supposed to run indefinately this is a pretty major problem. Thanks to StoykovK on the Althea discord for reporting!