Add fuzzing for event parsers
This patch adds fuzzing for Orchestrator event parsers, this isn't a
good fuzzing implementation but it does find bugs and provide stronger
garuntees that this coded will not crash.

Since submitting signatures is the only thing required to actually keep
the chain running an event that causes an error in this code is not
fatal. The bridge will simply wait until 2/3 of the validators deploy
an update to the orchestrator.

Now what is fatal is parsing an event incorrectly, these test will cause
that to happen if it's possible, but we don't have a good way to narrow
down valid versions from random data versus false positives.