Properly syncronize valset power threshold
In the previous patch "Hardcode power in Gravity.sol" we failed to
ensure that the Orchestrator and Gravity module where using exactly
2/3 * 2^32 rather than 66% of u32:MAX or some other slightly different
value.

This patch handles that, using the exact same value as the threshold
across all components.

This patch also adds a check required by "Ensure that validator set
updates have enough power" to make the relayer to print an error upon
encountering that condition.

It shouldn't be possible to encounter that in the first place given
that all new validator sets have no zero validators since the patch
"Do not include validators without an Eth addres in valsets". The only
way to generate such a validator set would be improper signing or
implementation error.