Validator sets will always have set addresses
As of the commit "Ensure that validator set updates have enough power"
it is no longer possible for a validator set from Cosmos to have an
address of 'None'. As of more recent commits validating Ethereum
addresses we can also be sure that there will not be any invalid
addresses in the Gravity Module store.

With this in mind we no longer need to represent Ethereum addresses as
Option<EthAddress> types, this reduces complexity across the board in a
few locations.