Prune actions docker cache every run
Github provides a 5gb cache limit per repo. This is more than big enough
for one set of build artifacts but not large enough to cache images
across multiple builds.

Lucky for us we only need to cache the artifacts within a single run,
happy-path-geth builds the images and then all the following builds
consume them. There won't be any cache hits across runs becuase of the
way we import the source code of the repo into the build. Doing this
busts our cache every time we change the code.