Upgrade Gravity.sol to use a struct for signatures
This patch adds a struct "Signature" which contains the v, r, and s
components of a validator signature. This is a more compact encoding and
removes the possibility of mismatched signature array lenghts.

Furthermore this reduces our total stack usage and allows us to declare
validator sets and signatures to be of type 'calldata' universally.

Total reduction in gas cost is about ~5% for validator set updates,
which are the most heavily impacted.