Update deep_space to simplify tx sending
This patch pulls in a new deep_space version that simplifies tx sending
and reduces boilerplate. In addition to reducing boilerplate we actually
estimate the gas usage of our Comsos transactions now.

This fixes scenarios where a large batch of validator set signatures
exceeds our hard coded gas usage, prevent submission and slashing a
validator.

Unfortunatley due to limitations in the Cosmos sdk our gas estimations
are not exact and instead are multiplied by two. So it's still possible
for us to have a perfectly valid and possible to submit transaction fail
because the now doubled gas requirements no longer fit in a single
block.