Do not accept zero address as Gravity contract location
I observed this as a bug during our recent testnet.

The gravity bridge contract address exists as a chain parameter that can
be set by governance, once set clients use this address by default and
no longer have to specify it on the command line.

Previously we detected that the contract address was not set as a
parameter when we found an empty string and failed to parse. But now
that we perform more type validation on the Gravity module end we are
instead returning hte totally valid zero address. Which passed parsing
and caused the orchestrator to go looking for contract events that it
will never find.

Even worse due to the design of oracle resync it will check the entire
chain history before giving an error, causing the uneducated user much
frustration and lost time.

This patch corrects the issue by checking aginst the zero address
before accepting the parameter value.