Wait for correct nonces in Unhalt bridge
This patch increases the robustness of the Uhalt bridge test by waiting
for the correct nonces, instead of erroring if they are not true right
as we check the condition.

We also remove a nonce check immediately following the false event. Due
to validator set updates and other background contract events we can
never really be sure of the correct offset without much more complicated
computation. It's best just to move onto checking that the bridge is
halted.