Bulk batching and test updates
The code was modified to handle the change in the proto-
files in order to allow for multiple batches to be validated at
the same time. Unit tests were updated to test the new change in struct
as the old definition would not work due to type issues.

Finally, the main loop function was modified to allow for bulk batch testing by
grouping multiple batches and logic calls which should prevent issues
with the blockchain being overwhelmed thus causing a halt in the chain.
The code was also modified to submit returns for multiple messages in
addition to handling multiple messages at once. This should theoretically
prevent attacks that attempt to overwhelm the blockchain by submitting
many logic calls and batches at once.