Display total supply on Ethereum for Cosmos assets
This patch modifies the ERC20 contract deployed by Gravity.sol in order
to represent Cosmos assets. We now override the totalSupply() function
in order to represent the total number of outstanding tokens on
Ethereum.

Previously this function would simply return uint256 max. This is
becuase we have these ERC20 contracts assign the maximum possible
blanace to the Gravity Bridge contract on creation. This allows us to
avoid managing any token minting logic in Gravity.sol.

The total supply on Ethereum is therefore uint256 max - the balance
currently left in the bridge. This is a more useful number than a simple
constant for users who might want to know about the supply of the token.

The ideal situation here would be to provide an accurate total supply
from the Cosmos side, but this is impractical for many reasons, mostly
revolving around how to get the updated value from the Cosmos to the
Ethereum side of the bridge efficiently and also managing those updates
in Gravity.sol.