Correct ERC20 metadata for Cosmos assets
This patch is focused around fixing the metadata for Cosmos assets when
represented on Ethereum and checking those fixes in the Rust tests.

Metadata for Cosmos assets is quite extensive. Tokens have a 'denom'
which is the string that primarily represents the token, but other names
can be provided for various quantities as well as a longform 'name'
field and a short symbol field.

For example
denom: ugraviton <= meaning micro-graviton
name: The Graviton token
symbol: GRAV
display: Graviton (display also includes how many decimals, so one
Graviton would be one million of the base denom ugraviton)

The name and symbol fields map directly onto the ERC20 name and symbol
fields. Previously we where checking that an adopted ERC20 matched the
display metadata element which is incorrect. Furthermore we where never
checking that these values where set correctly.

Also added while checking in the Rust tests that the ERC20 is properly
formed is a check to ensure the max suply is correct and that the
decimals are correct.