test_erc20_deposit transaction must enter before attestations
This is a minor cleanup patch that fixes a small ordering issue.

Attestations are created on the Cosmos chain by the validators once
their Orchestrator processes observe an event on Ethereum. Given these
constraints it obviously doesn't make sense to wait for the Ethereum
transcation *after* checking for the attestations.

I believe this is masking an error in send_to_cosmos and causing a
spurious test failure who's cause can't be directly observed.