Don't panic on attestation failure in test_erc20_deposit
Despite the fact that this testing function returns true or false
(modifed to a result in this patch) the attestation checking component
will panic on failure.

This patch corrects this oversight, one that was allowing unhalt bridge
to pass only in the case of strange race conditions. If everything
executed in the expected order the test would never pass becuase the
attestation would never be found.