No Orchestrator in deposit overflow
We don't need an orchestrator in this test, because we only submit false
events anyways having one running just opens up the chance of a race
condition around oracle submission ordering.

The only place where the orchestrator was used was for testing bridge
function before the start of the test. We can test this using the other
integration tests instead.