Fix nondeterministic genesis export for attestations
GetAttestationMapping should return ordered keys as well as the
mapping to make it easier for callers to iterate in a deterministic way.

All callers of GetAttestationMapping where performing the sort
themselves, except for the genesis file where failure to do so did not
cause any immediate errors in the tests.

This patch both improves the ordering in the genesis file and acts as a
general cleanup of repeated sorting logic.