Block SendToEth with zero address
OpenZeppelin forbids transfers to the zero address. So if a user where
to simply insert a transfer to zero into a batch they could cause the
entire batch to panic and fail to execute. Effectively dossing the token
in question until a chain upgrade could be deployed.

This provides protection for only one address, in general we'll need to
expand this protection.