Improve error messages around delegate keys
This patch resovles an issue where the error message for not having
delegate keys set encouraged users to register them, which isn't always
the correct thing to do.

If the user has already registered keys and simply not set them in this
orchestrator attempting another registration will almost definatley
create confusion about which keys are correct.

Also added in this patch are simple instructions for how to resolve
delegate keys issues, which will hopefully reduce the number of confused
users.