Detect uninitialized account
This adds a special case to the zero fees detection. A user has not been
assigned an account number when they have no tokens, therefore it's not
possible to query their base account. Since you need the account number
of a user to sign a valid transaction for them this precludes sending
transactions.

Since the Orchestrator can only function by sending transactions we
should exit in this case.