Fix nil pointer in gravity slashingkeeper
This patch resolves a nil pointer error in the Gravity slashing keeper
introduced with the airdrop proposal commit. Due to an oversight
app.slashingkeeper was nil in app.go. But this would not be noticed
until slashing happened, by default 10000 blocks after chain start.

This patch mades changes to app.go which will prevent these sorts of
accidents in the future. By using initialized variables instead of
references throughout the file. With this convention the compiler will
give us a proper warning if the variable is not initialized.

This patch also adds a new integration test designed to detect future
issues with the slashing logic by running slashing in our integration
tests. This is the only way to catch these sorts of errors since unit
tests do not use app.go in the first place.