Panic on unknown test type in test_runner
This patch improves a chain in the previous commit with the goal of
preventing the test runner from silently failing to the happy path test
when provided a unknown test type.

Since the happy path test is the 'default' it runs when no test type is
passed to the test runner. But the way this was handled caused us to run
the happy path test in any case where we could not parse the expected
test type.

This could result in us silently running the wrong test in CI, over and
over again and never knowing without looking carefully. This patch
resolves that issue while maintaing existing behavior with the happy
path test.