Forbid the creation of duplicate delegate keys
In a recent commit a new validation check was added to the integrity
checks for Ethereum validator sets that prevented duplicate eth
addresses as a general sign of corruption.

When this change was made the function SetOrchestratorAddress was not
updated to reflect this fact. Allowing anyone to come along, set a
duplicate key and halt the chain just a few blocks later when a new
validator set was created.

This patch resolves the issue by adding a check to
SetOrchestratorAddress and adding a check to the genesis file save/load
logic.