Prevent empty batch creation from panicing
This is a minor fix to a bug that doesn't actually have any effect on
chain functionality. When requesting a batch, if there where no
transactions to place in the batch the BuildOutgoingTXBatch function
would panic. This panic was recovered and had no effect on the chain
becuase it only occurred when no batch was going to be created either
way.

This patch simply fixes the panic and replaces it with a better error.