Ensure that there are no duplicates in send_ethereum_claims
This isn't strictly needed, the Ethereum contract should enforce
uniqueness and no two events should ever share an event nonce.

But there is value in adding a check here to make sure that that is
always ture. For example in some sort of Ethereum full node bug where
results from two forks are returned to us.