Handle more errors
This patch makes an attempt at handling all previously unhandled errors
in the Gravtiy module codebase.

Only a few of these are pure oversights, errors that where returned that
where not used or monitored. Most where intentionally ignored with the
assumption that the data was validated elsewhere. This is debatable, on
one hand a comparsion is a comparison it has an actual computation cost.
On the other hand errors existing becuase of spooky action at a
distance, failure elsewhere in the code is bad.

This patch does not go all the way in either direction, instead it
handles errors where the validation is 'further away' meaning not
performed in the same scope or at the same time as  as the
previously ignored error.