Orchestrator should not panic when locating an invalid signature
Previously I believed this error to be impossible, with geth validating
the signatures as they enter the chain only an rpc corruption error or a
signature of a type that Clarity doesn't support could cause this issue.
Yet we just saw it in produciton. For now we will demote this to an
error but it bears further investigation.