Validate address in confirm messages
Confirm messages, ValsetConfirm, BatchConfirm, and LogicCallConfirm are
all submitted with an attached Ethereum address. This is not actually
used in the validation on chain. We load the delegate key from the disk
and verify the signature against that.

But it is used by relayers to help index the message and determine who
should be attached to what signature. Therefore this data should be
validated like all the ohter data in the message.

This commit adds a test that confirms the submitted address is the same
as the delegate one in the store used to verify the signature preventing
potential attacks where a malicious validator submits a valid signature
with an incorrect attached address that prevents any relayer from
ordering the signatures properly to relay the resulting message.