Update to Clarity 0.5.0
This patch updates to Clarity 0.5.0 which removes the to_public_key()
function and replaces it with to_address(). This new version performs
more validation on the creation of the PrivateKey object, including
testing if a valid address can be derived from it. As part of that
process we cache the address in the PrivateKey object allowing us to
easily retrieve this data later.

This both reduces the number of unwraps in the code and also should
moderately improve performance as computing the address is a non-trivial
operation considering how slow bigint math is.