Use the higher level send_message for send_ethereum_claims
the goal of this patch is to use estimated gas values for
send_ethereum_claims, this is a very important function in the
orchestrator that actually submits oracle messages to the Cosmos
blockchain. If for some reason a large number of events caused the
resulting message to exceed it's hardcoded gas it would fail forever and
halt the bridge oracle.

The send_message function is higher level than send_transaction and
includes automated gas estimation on top of other ergonomic
improvements.

An additional advantage of this change is that transactions will not
fail if the orchestrator's validator is jailed. Instead they will fail
at the simulation level and never enter the chain.

The one downside of this change is that cosmos gas estimation is
somewhat inaccurate. So the effective fudge factor required to ensure
succees may result in a tx that could actually fit within a block being
rejected for requesting too much gas. This is of course still better
than the hardcoded gas that we currently use.