Actually compute amount of stake to delegate
Previously we just took a rough guess at how much we needed to delegate
in order to trigger a validator set update. This patch changes that and
introduces a utility function that computes the correct validator and
amount to delegate to. This should dramatically increase the reliability
of the VALSET_STRESS test.