Add slashing delegation test
This patch adds another CI test, the slashing delegation test that
validates the behavior of delegating and withdrawing rewards when
interacting with validators who have been slashed by the Gravity mdoule.
Currently this is the subject of a bug which prevents these delegators
from withdrawing their rewards. This test provides a reliable
reproduction of the bug, as well as coverage that will hopefully prevent
it from re-occuring once fixed.