Reduce ethereum signer failure cases
For the Orchestrator the ethereum signer is the only component that may
cause slashing if it fails. Therefore any error condition should be
looked at carefully.

This patch removes any dependence on Ethereum RPC requests for the
Ethereum signer, as it is not required to successfully create and submit
signatures. Furthermore an error where a single Cosmos gRPC request
failure may cause the loop to silently not run has also been corrected
and replaced with an appropriate retry system.