Fix go unit tests
In the commit "Remove and safeguard against app init errors" the staking
keeper cached in the GravityKeeper was replaced with a literal
StakingKeeper object rather than an interface. Several tests dependend
on this interface to provide a simple mock staking keeper.

Since we can no longer effectively use this mock staking keeper this
patch adds a new test environment creation function that allows us to
use the more complex 'CreateTestEnv()' with the arbitrary number of
validators and weights provided by the old lightweight mock staking
keeper.