Update reproducible image builder
The io package is only available in go 1.16 and later. The reproducible
builder was using go 1.15 resulting in an error. This patch uses a
newer version of the builder running go 1.17 which does not have this
issue.

The latest release tag is also updated to 1.1.0 since the changes to
slashing logic and hooks represent a breaking change.