Add governance features to gbt
This patch adds the ability to submit all of Gravity Bridge's custom
governance proposals to the gbt cli. As well as a query function that
allows users to interpret the airdrop proposal output locally.

This somewhat duplicates functions found in the gravity cli at this
point, this is required to submit the larger airdrop proposals.