Correct invalid deposits test
Someone sent a deposit to an Ethereum address on the live chain, since
the MsgSendToEth contains a freeform string there are lots of possible
ways it can be invalid.

The oracle component of the orchestrator has the goal of detecting
invalid deposit conditions and replacing them with an indication to the
Gravity Bridge module that this deposit should go into the community
pool. This keeps the oracle moving and prevents one bad deposit from
halting the bridge.

Sadly due to a significant oversight in the creation of the test the
invalid events test was merged with it's invalid deposits section
disabled. This patch restores it, corrects the regression that snuck
in and adds a new test case specific to what failed in production.