Add relay_every_valset relayer option
The primary goal of this change is to make it safe for a user to run the
relayer with the default settings without valset relaying exhausting all
their funds.

Previously there where two sets of validator set relaying behavior, one
where every valset was relayed and one where only profitable valsets
where relayed.

Relaying every valset is extremely wasteful in terms of gas and isn't
good default behavior for an altruistic user. Relaying only profitable
valsets is a valid choice, but will not keep the chain running correctly
if there are no rewards or the reward is too low.

This new default state, defined by having both relay_every_valset and
valset_relay_market as false in the config. Will only relay validator
sets when it becomes absolutely required for the bridge to continue
functioning. This should be extremely infrequent, with timing on the
order of several weeks.

This patch also contains a three other changes.

1) Orchestrator and Relayer configs are made 'copy' and references to
   them are removed throughout the relaying call stack
2) Valset relaying has been refactored and many comments and variable
   names touched up to improve readability and better define what each
   step is intended to do
3) valset_relaying.rs has been re-ordered so that a top to bottom
   reading of the file matches the call stack.