Get gravity_id from parameters
This patch removes a persisent an annoying error printed while running
the relayer. Previously we would grab the gravity_id from the contract
on Ethereum every iteration of the relayer main loop. This is prone to
spurious errors thanks to changes in the base fee between blocks.

This patch modifies the relayer main loop to retrieve this data from the
gravity bridge parameters store on Cosmos, which is not subject to this
error.

Additionally we do the operation only once instead of every loop
iteration. Since the gravity_id is immutable on Ethereum and can't be
changed on Cosmos without breaking the bridge.