Reduce relayer double submission by slowing down relaying
Gravity Bridge relayers have a fundamental race condition thanks to the
Ethereum mempool. Before submitted the TX the relayer will check the
state of the blockchain and then simulate the execution.

After checking that the submission is valid the relayer submits it's tx
and some time within the next minute a new block will be created
executing that transaction.

This is where our race conditon lies, the relayer is checking the
present state of the chain and so are many other relayers, each of them
queues up thier own tx for the next block and all but one will lose
money as their tx reverts shortly after starting.

Even worse since most of the gas cost of the transaction goes into it's
arguments the relayer pays a high price for this failure.

What this patch does is reduce the probability of this condition by
dramatically reducing frequency at which the Ethereum relayer runs.

Assuming users start their nodes at an evenly distributed period and a
fixed block time, neither of which are prefectly true we come up with
this approximate equation.

expected number of collisions = (num relayers - 1 ) * (block time / loop
time)

For a 17 second loop time and two relayers we expect .88 collisions

For a 600 second loop and two relayers we expect .025 collisions

For a 600 second loop and one hundred relayers we expect 2.5 collisions

A proper solution for this issue will detect collisions and slow
down/speed up dynamicly. This is made more complicated by batches, where
this may have to be done individually across all token types.

For now we provide the single knob, loop time for relayers to adjust
their risk tolerence.