Better batch gas display
This patch adds a seemingly simple feature, better display of batches
that includes values with exchange rates as well as token symbols. This
requires not only a lot of requests to the ETH node but also a
significant amount of number conversion.

This lead me to realize that there was duplication between number
conversion required in GBT and how it was being done in the relayer,
resulting in a minor refactor to reduce this duplication.