BridgeContractAddress -> BridgeEthereumAddress
This is a minor naming oversight where the field in genesis is named
BridgeEthereumAddress but the name registered with the paramater store
was BridgeContractAddress. This results in some confusion when using
governance to update the parameter.

This patch while very minor is consensus breaking, so it shouldn't be
merged until we're getting closer to an upgrade.