Add integration test for GravityERC721.sol
This patch adds an integration test for GravityERC721.sol. Since the Go
components are not yet done for NFT support all that is tested is the
deployment of the GravityERC721.sol, deployment of ERC721 test contracts
and the basic sendERC721ToCosmos contract call.

The test confirms that an event is emitted and the NFT is transferred.
Next steps are to begin work on the Gravity Module side before returning to the
orchestrator to complete this test.

Commit message by Justin Kilpatrick <justin@althea.net>