Fix double call of RegisterProposalTypes
A recent commit "register proposal types for the cli" caused the gravity
node to panic on startup due to double registration of proposal types.
Being overconfident that it was a simple fixed I pushed this straight to
master and broke the tests there.

This patch fixes the issue, which somehow invovles App.go calling all
the CLI command code and duplicating the proposal registration. We need
to register proposals in the CLI for the case where we are not running a
full node and App.go does not register the proposal type. The ultimate
soltuion is a check to prevent double registration.