Add TX count to the BatchFees struct (#50)
Knowing how many transactions are waiting to be relayed allows us to estimate how much it's going to cost to relay the hypothetical batch. This allows orchestrators to better decide when to request a batch but more importantly, allows the UI to suggest how much the user should put in bridge fees.