Update app.go to have better consistency
We made an effort in app.go to initialize structs and then use
references to those structs in the gravity app. I changed the struct
initialization to ensure we have no nil values, removed use of
app.struct in favor of the raw struct value use, and removed some
useless authz initialization.