Fix ValsetConfirm iteration
The function GetValsetConfirms breaks when the validator set nonce
exceeds the value of 128. This is because types/key.go was modified to
use strings for indexing but this function was not modified to perform
the same conversion when deciding it's prefix range.

This has the unfortunate effect of halting the chain, for validator sets
greater than 128 becuase the confirms will not be successfuly located in
abci.go when valset slashing is performed. Resutling in slashing of all
the validators at the same block.

This patch corrects the issue for GetValsetConfirm by stringifying the
nonce in the same way as types/key.go

But the existance of this issue highlights a larger need to further
restructure key.go to provide prefixed keys so that modifications are
more localized and the addition of more unit tests related to getters
and setters.

I have found IterateLogicConfirmByInvalidationIDAndNonce suffers from
the same issue and generally question the existance of
ConvertByteArrToString in key.go.