More comprehensive airdrop integration test
This patch expands the airdrop integration test to handle extra
conditions. Specifically relating to airdrops larger than u64 max and to
ensure the total supply and amount changed in the community pool are
correct.

This was achived by disabling inflation and refactoring the airdrop
proposal utility functions to be easily repeatable.