Add invariant checking to CreateTestEnv
This patch adds a simple invariant checking function to the TestInput
struct and runs it at the end of the unit test environment setup
function CreateTestEnv.

The test environment is modified to comply with all invariants and on
test required changes in response to the invariant fixes.

Instead of using the crisisKeeper invariants assertion we manually call
the invariants functions for each keeper we declare in our test
environment. While this isn't ideal in terms of lines of code it's much
easer to setup than adding the crisis keeper to GravityKeeper and
dealing with initializing all the used modules.