Fix airdrop proposal invariant violation
The Airdrop proposal handler summed the total amount of tokens required
into a u64 max, this could have resulted in a proposal not subtracting
the correct amount from the community pool if greater than u64 max
tokens where being aidropped.

The net effect here would be to mint tokens, something the airdrop
proposal is not intended to do.

In theory this would also bypass the check determining if the community
pool had a sufficient number of tokens, but that would simply fail later
during the transfer phase and revert the proposal anyways, rendering it
an unexploitable outcome.