Double test governance voting period
This doubles the time for governance votes to execute in the integration
tests from 60 seconds to 120 seconds. This is required for the slow test
runners in Github actions combined with the new Airdrop test trying to
do back to back airdrops. The chain is simply not running fast enough
in this case to get the validators votes in before 60 seconds elapses.

This change should be reconsidred or reversed if the minimum spec of the
CI runners improves.

Note, this patch also fixes one last spot where the proposal voting
duration was hardcoded.