Prevent spurious warning when auto-requesting batches
This patch adds a check for when the auto batch request tx fails becuase
the new batch would not be more proftiable. This is intended behavior
and our profitability check does not take this into account.

When the orchestrator decides to request a batch it only checks if there
are any transactions with fees in the tx pool. It does not check if
there is already a more profitable batch of this type waiting. But the
actual batch creation logic does perform this check to prevent the spam
creation of batches.

Since it's free to make this mistake (tx fails in simulation) we can
simply catch the error and mute it rather than expanding the
orchestrator logic to match the internal batch creation logic.