Do not panic on confirms from validators that no longer exist
Previously when looking up a validators orchestrator address during
slashing we used GetOrchestratorValidator() which queried the validator
set and returned the validator entry from the staking module.

When a validator unbonds completely it will no longer be in this list.
So when we go to query the validators for every confirmation message
containing an ethereum signature the lookup will fail and the code will
panic. Since this particular panic is occuring in the endlbocker this
means all validators will panic simultaneously and the chain will halt.

This patch resolves the issue by no longer referencing the staking
module and simply looking up the mapping from valoper to orchestrator
address in the Gravity store. This is a more robust solution in general
as we never used the info from the staking module.

Commit message by Justin Kilpatrick