Add GravityERC721.sol (#48)
This commit adds a new solidity contract designed to hold NFT assets for Gravity Bridge. Since the original Gravity contract is immutable this contract uses the arbitrary logic functionality. Where a call is passed through Gravity.sol to this contract, which only allows itself to be called from Gravity.sol's address as set in the constructor.

Included is a contract with batched withdraws and event based deposits as well as unit test mirroring Gravity.sol's ERC20 version of the same functionality as closely as possible.