Resolve attestation order race conditon (#124)
This patch resolves a race condition around the unhalt bridge governance vote. Previously if a new event was submitted during the time the bridge was waiting for the nonces to be reset the Orchestrator code would not correctly detect the rollback and would remain stuck trying to submit an out of order event.

Also included are improvements to the unhalt bridge test that reproduce this behavior and confirm the fix. 

Co-authored-by: dusan.maksimovic@ethernal.tech <dusan.maksimovic@ethernal.tech>