Improve gbt client deploy-erc20 error behavior
Previously when deploying an erc20 representation gbt would only wait
for 100 seconds before declaring the attempt a failure.

Since the claims process requires 2/3 of the voting power worth of
validators to successfully send a tx this is not long enough in a
production environment and would result in a spurious error.

This patch dramatically improves the error behavior by both waiting
longer and inspecting the recent attestation history to determine what
may be wrong and present a message to the user that is accurate and not
confusing.

During development I noticed that the attestion query filtering features
do not seem to be working properly, I've opened an issue for this.